+++ go/tasks/plugins/webapi/agent/config_test.go
@@ -1,4 +1,4 @@
-package grpc
+package agent
 
 import (
 	"testing"
